+++ src/modules/sophia/components/index.ts
@@ -1,2 +1,6 @@
 export { EntityList } from './EntityList';
 export { EntityListItem } from './EntityListItem';
+export { EntityDetail } from './EntityDetail';
+export { EntityDetailEmpty } from './EntityDetailEmpty';
+export { EntityDetailHeader } from './EntityDetailHeader';
+export { EntityDetailContent } from './EntityDetailContent';